Refactor en la autenticacion: Ahora se usa api-key

+++ pystock/app/routes/product_controller.py
@@ -1,5 +1,6 @@
 from flask import Blueprint, jsonify, request
 from pystock.app.services.product_service import *
+from pystock.app.config import API_KEY
 
 
 product_controller = Blueprint('product_controller', __name__)
@@ -7,6 +8,10 @@ product_controller = Blueprint('product_controller', __name__)
 
 @product_controller.route('/setProduct', methods=["POST"])
 def set_product():
+    auth = request.headers.get("X-Api-Key")
+    if auth != API_KEY:
+        return jsonify({"message": "ERROR: Unauthorized"}), 401
+
     credentials = request.json
     cod = credentials["cod"]
     name = credentials["name"]
@@ -23,6 +28,10 @@ def set_product():
 
 @product_controller.route('/editProduct/<product_id>', methods=["PUT"])
 def edit_product(product_id):
+    auth = request.headers.get("X-Api-Key")
+    if auth != API_KEY:
+        return jsonify({"message": "ERROR: Unauthorized"}), 401
+
     credentials = request.json
     cod = credentials["cod"]
     name = credentials["name"]
@@ -36,10 +45,18 @@ def edit_product(product_id):
 
 @product_controller.route('/getProducts')
 def products():
+    auth = request.headers.get("X-Api-Key")
+    if auth != API_KEY:
+        return jsonify({"message": "ERROR: Unauthorized"}), 401
+
     return jsonify(get_products()), 200
 
 
 @product_controller.route('/deleteProduct/<product_id>', methods=["DELETE"])
 def delete_a_product(product_id):
+    auth = request.headers.get("X-Api-Key")
+    if auth != API_KEY:
+        return jsonify({"message": "ERROR: Unauthorized"}), 401
+
     delete_product(product_id)
     return "Deleted", 204
