+++ pystock/app/routes/user_controller.py
@@ -2,7 +2,7 @@ from flask import Blueprint, jsonify, request
 
 from pystock.app.utils.login_manager import check_login
 from pystock.app.services.user_service import get_id, get_users, insert_user, user_exist, delete_user
-from pystock.app.utils.token_manager import generate_token
+from pystock.app.config import API_KEY
 
 
 user_controller = Blueprint('user_controller', __name__)
@@ -17,16 +17,24 @@ def login():
     if not check_login(email, password):
         return "User not fond", 500
 
-    return {"id": get_id(email), "token": generate_token()}
+    return {"id": get_id(email), "apikey": API_KEY}
 
 
 @user_controller.route('/getUsers')
 def users():
+    auth = request.headers.get("X-Api-Key")
+    if auth != API_KEY:
+        return jsonify({"message": "ERROR: Unauthorized"}), 401
+
     return jsonify(get_users()), 200
 
 
 @user_controller.route('/setUser', methods=["POST"])
 def create_user():
+    auth = request.headers.get("X-Api-Key")
+    if auth != API_KEY:
+        return jsonify({"message": "ERROR: Unauthorized"}), 401
+
     credentials = request.json
     email = credentials["email"]
     password = credentials["password"]
@@ -39,5 +47,9 @@ def create_user():
 
 @user_controller.route('/deleteUser/<user_id>', methods=["DELETE"])
 def delete_a_user(user_id):
+    auth = request.headers.get("X-Api-Key")
+    if auth != API_KEY:
+        return jsonify({"message": "ERROR: Unauthorized"}), 401
+
     delete_user(user_id)
     return "Deleted", 204
